Added a basic LSTM model on DIC India
The basic LSTM model has been done on the stock prices of DIC India. I haven't followed the timeframe we're looking for and even the company is different but this is just a template which we can use once the preprocessing for the Alphabet data has been done. I've also attached the .csv file in case you need to look at the data.