Updated stockpricepred.py to near full-functionality
I have updated the Streamlit Python script, it is almost done and is showing forecasts for one week into the future for both companies. However, there are some dimensions errors I'm getting in the plotting of the graphs for the rest of the 2-5 week windows. That has to be fixed from the model design side. So once it's fixed, we will need to do model.save again into the appropriate model names as used in this script.